+++ src/Pim/MultiVendorProduct/MultiVendorProduct.php
@@ -72,6 +72,9 @@ final class MultiVendorProduct
     /** @var array */
     private $imageIds;
 
+    /** @var null|MultiVendorProductVideo */
+    private $video;
+
     /**
      * MultiVendorProduct constructor.
      *
@@ -95,6 +98,7 @@ final class MultiVendorProduct
         $this->freeAttributes = $data['freeAttributes'] ?? null;
         $this->imageIds = $data['imageIds'] ?? null;
         $this->attributes = $data['attributes'] ?? null;
+        $this->video = isset($data['video']) ? new MultiVendorProductVideo($data['video']) : null;
     }
 
     /**
@@ -401,6 +405,26 @@ final class MultiVendorProduct
     public function setImageIds(?array $imageIds): self
     {
         $this->imageIds = $imageIds;
+
+        return $this;
+    }
+
+    /**
+     * @return MultiVendorProductVideo|null
+     */
+    public function getVideo(): ?MultiVendorProductVideo
+    {
+        return $this->video;
+    }
+
+    /**
+     * @param MultiVendorProductVideo|null $video
+     *
+     * @return MultiVendorProduct
+     */
+    public function setVideo(?MultiVendorProductVideo $video): self
+    {
+        $this->video = $video;
 
         return $this;
     }